add delete route

+++ backend/controllers/stuff.js
@@ -1,4 +1,5 @@
 const Thing = require('../models/Thing');
+const fs = require('fs');
 
 exports.createThing = (req, res) => {
     const thingObject = JSON.parse(req.body.thing);
@@ -23,8 +24,6 @@ exports.modifyThing = (req, res) => {
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     } : { ...req.body };
 
-    console.log(thingObject);
-
     delete thingObject._userId;
     Thing.findOne({ _id: req.params.id })
     .then((thing) => {
@@ -40,9 +39,20 @@ exports.modifyThing = (req, res) => {
 };
 
 exports.deleteThing = (req, res) => {
-    Thing.deleteOne({ _id: req.params.id })
-    .then(res.status(200).json({ message: 'Objet supprimé !'}))
-    .catch(error => res.status(400).json({ error }));
+    Thing.findOne({ _id: req.params.id })
+    .then(thing => {
+        if (thing.userId != req.auth.userId) {
+            res.status(401).json({ message: 'Not authorized' });
+        } else {
+            const filename = thing.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+                Thing.deleteOne({ _id: req.params.id})
+                .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
+                .catch(error => res.status(401).json({ error }));
+            })
+        }
+    })
+    .catch(error => res.status(500).json({ error }));
 };
 
 exports.getAllStuff = (req, res) => {
